Write message to database in data handler
